+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/base/models/PostId.kt
@@ -1,0 +1,3 @@
+package com.github.insanusmokrassar.AutoPostTelegramBot.base.models
+
+typealias PostId = Int