+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/choosers/MostRatedRandomChooser.kt
@@ -1,7 +1,9 @@
 package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.choosers
 
+import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.PostId
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
+import org.joda.time.DateTime
 import java.util.*
 
 @Serializable
@@ -9,8 +11,8 @@ class MostRatedRandomChooser : RateChooser() {
     @Transient
     private val random = Random()
 
-    override fun triggerChoose(): Collection<Int> {
-        return postsLikesTable ?.getMostRated() ?.let {
+    override fun triggerChoose(time: DateTime, exceptions: List<PostId>): Collection<Int> {
+        return postsLikesTable ?.getMostRated() ?.minus(exceptions) ?.let {
             if (it.isEmpty()) {
                 it
             } else {